[Reland] Handle appearance value changes during <select> showPicker.
When the appearance value changes as the result of the picker
opening (e.g. via `::picker(select):open {appearance:auto}` or
similar), we close the picker to avoid any oscillation or
nonsensical behavior. This CL implements this change.

The original CL [1] was reverted due to a strange bot failure [2]
where the new set of promise_tests somehow fail because
blink_wpt_tests is expecting an expectation file with a single blank
line. I added a "Failure" expectation to hopefully avoid that, in
Patchset 3.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/6094831
[2] https://ci.chromium.org/ui/p/chromium/builders/ci/mac11-arm64-rel-tests/50289/overview

Fixed: 376097114,370536090,367426156
Bug: 384394713
Change-Id: I2ab4e43db3de4b7125b564c366fe86bdf9f85cc4